[cluster] Removed some files and command that are no longer needed. rhbz #1083656
Removed some of the files and commands that are no longer needed on RHEL6+.
In addition, the gfs_lockdump option did not capture gfs2 lock dumps.
The option isnow called gfs2_lockdump and will mount the /sys/kernel/debug
directory and the gfs2 lockdumps will be copied from that mount point.

In addition added a couple files that are needed for pacemaker/dlm.

Signed-off-by: Shane Bradley <sbradley@redhat.com>